get video url at time of data push
if you use a single iframe to support multiple videos on the page and swap out the src attribute on the iframe, the tracking code doesn't capture the actual URL when pushing data.  Instead you get an error like this:

```
Failed to execute 'postMessage' on 'DOMWindow': The target origin provided ('https://initial_iframe_url') does not match the recipient window's origin ('https://player.vimeo.com').
```

This changes the script so that it obtains the URL of the iframe just prior to pushing data so that it matches the expected origin