jQuery load robustness + allow event label modification
This change makes the script load a bit more robustly when jQuery
may not be initialised and will label events with both the URL
of the vimeo video and the title of the iframe. This can make it
easier to look at the analytics and see what videos were being
viewed.